requirements.txt
elasticsearch==8.11.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
rich==13.7.0 